move representation_creation utils functions
