+++ bia-ingest/test/test_conversion_util_funcs.py
@@ -1,4 +1,4 @@
-from bia_ingest.ingest import utils
+from bia_ingest.representation_creation import utils
 
 
 def test_merge_dicts():
